Update Huffman_Encoding.cpp
compress() and decompress() functions now operate with the ascii_letters array stored in the compressed file. In addition, encoded file stores the number of valid bytes in the last encoded byte (avoiding extra 0's at the end). This number is added to the end of the encoded file. The decompress function handles the addition of array_size, ascii_letters array, and valid_bytes accordingly and only reads the encoding section of the file